requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
httpx==0.25.1
pydantic==2.5.0
pydantic-settings==2.1.0
slowapi==0.1.9
psycopg2-binary==2.9.9